A property is only valid if all implications hold.
If any implication fails, this means the data isn't valid for a given property
and as such, the result should be discarded. The previous version ran a
property if any implication held, which could lead to weird behaviour in case
of more than one implication.

e.g.:

    for-all Array(Int)
      .given (.length > 0)
      .given (.every (-> typeof it is 'string'))
      .satisfy (...)

Would run the property none-the-less, even though the array clearly has no
string elements.